requirements.txt
# Multi-Modal AI Platform Dependencies (Student Edition)

# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
accelerate>=0.20.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
moviepy>=1.0.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Database & Vector Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
weaviate-client>=3.25.0

# Cloud Storage
google-cloud-storage>=2.10.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development & Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Monitoring & Logging
structlog>=23.1.0
prometheus-client>=0.17.0 